Removed load settings from constructor and moved to initialize in AbstractCommand.
Loading settings initialize and not in the constructor allows the `list` command to work properly without the end
user actually needing to create a config file.
Bonus: when the Settings class throws a runtime exception it shows a clean error message to the user. In the
previous implementation we did not have clean error messages.